+++ dep0/lib/00_core/src/cxx23.cpp
@@ -1,0 +1,1 @@
+#include "dep0/cxx23.hpp"
